lib/ukboot: Provide `wait_for_debug` command line option
Introduces the option to busy wait during booting until a debugger
is connected and sets a variable to 0. This can be helpful if
such a behavior cannot be achieved with the execution environment,
like the hypervisor.
The busy loop is only enabled when the kernel command line option
  `boot.wait_for_debug=1`
was given. The busy loop is placed to the earliest place possible,
after command line arguments are parsed.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>