plat/common/paging: Lazy PROT_NONE allocation
Many applications allocate mmap regions with PROT_NONE and later
change protections to reserve virtual memory space upfront without
actually allocating physical memory (or page tables). This commit
introduces a hack to mimic this behavior with the prototype
posix-mmap library that does not support this.

The commit must be removed as soon as the posix-mmap
library is in a clean and mature state and supports this behavior
without the hack.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>